Extract update functions
